added check for OverFlow Exception

+++ WeeklyProject2/FibonacciClass.cs
@@ -4,20 +4,20 @@ namespace WeeklyProject2
 {
     class FibonacciClass
     {
-        public static int Fibonacci(int secondNumber)
+        public static ulong Fibonacci(int secondNumber)
         {  
             if (secondNumber == 1 || secondNumber == 2)
             {
                 return 1;
             }
-            int prev_number = 0, curr_number = 1, next_number = 0;
-            for (int index = 1; index < secondNumber; index++)
+            ulong prev_number = 0, curr_number = 1, next_number = 0;
+            for (ulong index = 1; index < (ulong)secondNumber; index++)
             {
                 next_number = prev_number + curr_number;
                 prev_number = curr_number;
                 curr_number = next_number;
             }
-            return next_number;
+                return next_number;           
         }
 
         public static bool FibonacciSequence(int firstNumber)
